+++ backend/api/schemas/habit_schema.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+from typing import Optional
+from datetime import date, datetime
+
+class Habit(BaseModel):
+    id: int
+    description: str
