+++ backend/api/schemas/user_habit_schema.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+from datetime import date
+
+class UserHabit(BaseModel):
+    id: int
+    start_date: date
+    is_active: bool
+    user_id: int
+    habit_id: int
