+++ backend/api/services/get_users.py
@@ -1,0 +1,2 @@
+def get_all_users(users):
+    return users
