+++ backend/main.py
@@ -1,76 +1,67 @@
-from fastapi import FastAPI , HTTPException, Depends, status
-from models import User, Habit, UserHabit
-from datetime import date, datetime
-from database.session import db_dependency
+from fastapi import FastAPI
+from api.routes import user
 
 
 app = FastAPI()
 
-test_users = [
-    User(
-        id=1,
-        username='erinforrest',
-        password_hash=1234,
-        date_created=date.today(),
-        last_login=datetime.now()
-    )
-]
+app.include_router(user.router)
 
-test_habits = [
-    Habit(
-        id=1,
-        description='Take the Subway'
-    )
-]
 
-test_user_habits = [
-    UserHabit(
-        id = 1,
-        start_date = date.today(),
-        is_active = True,
-        user_id = 1,
-        habit_id = 1
-    )
-]
+# test_habits = [
+#     Habit(
+#         id=1,
+#         description='Take the Subway'
+#     )
+# ]
 
-@app.get("/users")
-async def fetch_users():
-    """
-    Args:
-        None
+# test_user_habits = [
+#     UserHabit(
+#         id = 1,
+#         start_date = date.today(),
+#         is_active = True,
+#         user_id = 1,
+#         habit_id = 1
+#     )
+# ]
 
-    Raises:
+# @app.get("/users")
+# async def fetch_users():
+#     """
+#     Args:
+#         None
 
-    Returns:
-        The list of all users in teh test_users list
-    """
-    return test_users
+#     Raises:
 
-@app.get("/habits/{user_id}")
-async def fetch_user_habits(user_id: int):
-    """
-    Args:
-        user_id: passed in through the get URL
+#     Returns:
+#         The list of all users in the test_users list
+#     """
+#     return test_users
 
-    Raises:
-        404 User not found: if the user_id in the get URL does not exist
-        404 No habits found for this user: if there are no habits that match the given user_id
+# @app.get("/habits/{user_id}")
+# async def fetch_user_habits(user_id: int):
+#     """
+#     Args:
+#         user_id: passed in through the get URL
+
+#     Raises:
+#         404 User not found: if the user_id in the get URL does not exist
+#         404 No habits found for this user: if there are no habits that match the given user_id
         
-    Returns:
-        A list of all the habits for a specific user_id
-    """
-    # First, check if the user exists
-    user = next((user for user in test_users if user.id == user_id), None)
-    if not user:
-        raise HTTPException(status_code=404, detail="User not found")
+#     Returns:
+#         A list of all the habits for a specific user_id
+#     """
+#     # First, check if the user exists
+#     user = next((user for user in test_users if user.id == user_id), None)
+#     if not user:
+#         raise HTTPException(status_code=404, detail="User not found")
 
-    # Find all UserHabit entries for this user
-    user_habits = [user_habit for user_habit in test_user_habits if user_habit.user_id == user_id]
+#     # Find all UserHabit entries for this user
+#     user_habits = [user_habit for user_habit in test_user_habits if user_habit.user_id == user_id]
     
-    if not user_habits:
-        raise HTTPException(status_code=404, detail="No habits found for this user")
+#     if not user_habits:
+#         raise HTTPException(status_code=404, detail="No habits found for this user")
 
-    # Get the actual Habit objects
-    habits = [(habit for habit in test_habits if habit.id == user_habit.habit_id) for user_habit in user_habits]
+#     # Get the actual Habit objects
+#     habits = [(habit for habit in test_habits if habit.id == user_habit.habit_id) for user_habit in user_habits]
 
-    return habits
+#     return habits
